+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- convertKeyboardLayoutToCharacterKeyCodeMap keep all mapping to a character
+
 ## [0.0.52](https://github.com/andy23512/tangent-cc-lib/releases/tag/0.0.52)
 
 - add CCLite device and device layout